+++ server-adapters/resteasy-undertow/src/main/java/org/jboss/resteasy/plugins/server/undertow/UndertowJaxrsServer.java
@@ -9,6 +9,7 @@ import io.undertow.servlet.api.ServletInfo;
 
 import org.jboss.resteasy.util.PortProvider;
 import org.jboss.resteasy.plugins.server.servlet.HttpServlet30Dispatcher;
+import org.jboss.resteasy.plugins.server.servlet.servlet3.HttpServlet30Dispatcher;
 import org.jboss.resteasy.spi.ResteasyDeployment;
 
 import javax.servlet.ServletException;
