+++ testsuite/integration-tests/src/test/resources/org/jboss/resteasy/test/asynch/JaxrsAsyncServletWeb.xml
@@ -5,7 +5,7 @@
 <web-app>
     <servlet>
         <servlet-name>Resteasy</servlet-name>
-        <servlet-class>org.jboss.resteasy.plugins.server.servlet.HttpServlet30Dispatcher</servlet-class>
+        <servlet-class>org.jboss.resteasy.plugins.server.servlet.servlet3.HttpServlet30Dispatcher</servlet-class>
         <init-param>
             <param-name>javax.ws.rs.Application</param-name>
             <param-value>org.jboss.resteasy.test.asynch.resource.JaxrsAsyncServletApp</param-value>
